Add files via upload
An arithmetic array is an array that contains at least two integers and the differences between consecutive integers are equal. 
for example- [9,10],[3,3,3] and [9,7,5,3] are arithmetic arrays while [1,3,3,7],[2,1,2]
and [1,2,4] are not arithmetic arrays.

Ques-Sarasvati has an array of N non-negative integers. The i-th integer of the array is A. She wants to choose a contiguous arithmetic subarray from her array that has the maximum length. Please help her to determine the length of the longest contiguous arithmetic subarray.

Input:
The first line of the input gives the number of test cases, T. T test cases follow. Each test case begins with a line containing the integer N. The second line contains N integers. The i-th integer is A.

Output:
For each test case, output one line containing Case #x: y, where x is the test case number (starting from 1) and y is the length of the longest contiguous arithmetic subarray.

Intuition and Approach:

• Loop over the array and find the answer.

Maintain the following variables:

1. Previous Common Difference (pd)
2. Current Arithmetic Subarray length (curr)
3. Max arithmetic subarray length (ans)